Keymap - removed momentary scroll hold from right click
The default AdeptBLE keymap has right click tap and momentary drag
scroll on the same key which meant it was impossible to use the mouse to
move window borders in Hyprland. Since momentary scroll is still on the
button below right click, I'm just making right click a normal mouse
keypress with no hold function on the default layer.